requirements.txt
torch>=1.13.0
torchvision>=0.15.0
ultralytics>=8.0.0
numpy>=1.22.2
opencv-python>=4.7.0
PyYAML>=6.0
tqdm>=4.65.0
matplotlib>=3.7.1
seaborn>=0.12.2
scikit-learn
# For Colab/Google Drive support
# gdown 